Refactor code structure for danger-kotlin-library
