+++ danger-kotlin-library/src/main/kotlin/systems/danger/kotlin/KtxBitBucketServer.kt
@@ -1,0 +1,5 @@
+package systems.danger.kotlin
+
+// extensions over [BitBucketServer] object
+
+// No extensions implemented
