+++ danger-kotlin-library/src/main/kotlin/systems/danger/kotlin/KtxGitHub.kt
@@ -1,0 +1,3 @@
+package systems.danger.kotlin
+
+// extensions over [GitHub] object
