+++ danger-kotlin-library/src/main/kotlin/systems/danger/kotlin/KtxGitLab.kt
@@ -1,0 +1,3 @@
+package systems.danger.kotlin
+
+// extensions over [GitLab] object
