+++ danger-kotlin-library/src/main/kotlin/systems/danger/kotlin/MainPlugins.kt
@@ -1,0 +1,25 @@
+package systems.danger.kotlin
+
+import systems.danger.kotlin.sdk.DangerContext
+import systems.danger.kotlin.sdk.DangerPlugin
+
+/**
+ * R
+ */
+object register {
+    internal var dangerPlugins = mutableListOf<DangerPlugin>()
+
+    infix fun plugin(plugin: DangerPlugin) {
+        dangerPlugins.add(plugin)
+    }
+
+    fun plugins(vararg pluginArgs: DangerPlugin) {
+        dangerPlugins.addAll(pluginArgs)
+    }
+}
+
+inline fun register(block: register.() -> Unit) = register.run(block)
+
+internal fun DangerPlugin.withContext(dangerContext: DangerContext) {
+    context = dangerContext
+}
