+++ danger-kotlin-library/src/main/kotlin/systems/danger/kotlin/models/danger/DangerResults.kt
@@ -1,14 +1,14 @@
 @file:UseSerializers(ViolationSerializer::class)
 
-package systems.danger.kotlin
+package systems.danger.kotlin.models.danger
 
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.UseSerializers
 import systems.danger.kotlin.sdk.Violation
-import systems.danger.kotlin.serializers.ViolationSerializer
+import systems.danger.kotlin.models.serializers.ViolationSerializer
 
 @Serializable
-data class Meta(
+internal data class Meta(
     val runtimeName: String = "Danger Kotlin",
     val runtimeHref: String = "https://danger.systems"
 )
