+++ danger-kotlin-library/src/test/kotlin/systems/danger/kotlin/UtilsTests.kt
@@ -3,6 +3,7 @@ package systems.danger.kotlin
 import kotlinx.serialization.decodeFromString
 import org.junit.Assert
 import org.junit.Test
+import systems.danger.kotlin.models.danger.DSL
 import systems.danger.kotlin.utils.TestUtils.JSONFiles
 import systems.danger.kotlin.utils.TestUtils
 import java.io.File
