+++ danger-kotlin-library/src/test/kotlin/systems/danger/kotlin/models/bitbucket/BitBucketServerParsingTests.kt
@@ -1,8 +1,9 @@
-package systems.danger.kotlin
+package systems.danger.kotlin.models.bitbucket
 
 import kotlinx.serialization.decodeFromString
 import org.junit.Assert.assertEquals
 import org.junit.Test
+import systems.danger.kotlin.models.danger.DSL
 import systems.danger.kotlin.utils.TestUtils.JSONFiles
 import systems.danger.kotlin.utils.TestUtils
